refactor(api): extract checkBoardAccess into shared boardAccess helper
- Lift the inline helper from boardCommentsController into apps/api/routes/boards/boardAccess.ts
- Mirrors the apps/api/routes/docs/documentAccess.ts convention
- No behavior change; enables reuse for the upcoming assignable-members endpoint
